cleaning up variable names
