+++ freemocap/__init__.py
@@ -25,6 +25,7 @@ logger = logging.getLogger(__name__)
 logger.info(f"Initializing {__package_name__} package, version: {__version__}, from file: {__file__}")
 
 
+# breadcrumbs for legacy/deprecated 'pre-alpha` entry point
 def RunMe(*args, **kwargs):
     logger.info(
         "User tried using `pre-alpha` entry point (`import freemocap: freemocap.RunMe() - displaying friendly message then re-directing to `freemocap.__main__:main()` entry point")
