+++ freemocap/parameter_info_models/recording_info_model.py
@@ -173,7 +173,7 @@ class RecordingFolderStatusChecker:
         try:
             test_image_tracking_data_shape(
                 synchronized_video_folder_path=self.recording_info_model.synchronized_videos_folder_path,
-                image_tracking_data_file_name=self.recording_info_model.mediapipe_2d_data_npy_file_path,
+                image_tracking_data_file_path=self.recording_info_model.mediapipe_2d_data_npy_file_path,
             )
 
             return True
@@ -185,7 +185,7 @@ class RecordingFolderStatusChecker:
             test_mediapipe_skeleton_data_shape(
                 synchronized_video_folder_path=self.recording_info_model.synchronized_videos_folder_path,
                 raw_skeleton_npy_file_path=self.recording_info_model.mediapipe_3d_data_npy_file_path,
-                reprojection_error_file_name=self.recording_info_model.mediapipe_reprojection_error_data_npy_file_path,
+                reprojection_error_file_path=self.recording_info_model.mediapipe_reprojection_error_data_npy_file_path,
             )
             return True
         except AssertionError as e:
