+++ freemocap/system/logging/configure_logging.py
@@ -13,6 +13,7 @@ format_string = "[%(asctime)s.%(msecs)04d] [%(levelname)8s] [%(name)s] [%(funcNa
 
 default_logging_formatter = logging.Formatter(fmt=format_string, datefmt="%Y-%m-%d %H:%M:%S")
 
+logging.basicConfig(level=logging.WARNING)
 
 def get_logging_handlers():
     dictConfig(DEFAULT_LOGGING)
