+++ freemocap/tests/conftest.py
@@ -5,7 +5,7 @@ from freemocap.utilities.download_sample_data import (
 )
 from freemocap.system.paths_and_filenames.path_getters import get_output_data_folder_path, \
     get_synchronized_videos_folder_path, get_raw_skeleton_npy_file_name, get_total_body_center_of_mass_file_path, \
-    get_image_tracking_data_file_name, get_reprojection_error_file_name
+    get_image_tracking_data_file_name, get_reprojection_error_file_path
 from freemocap.core_processes.process_motion_capture_videos.process_recording_headless import process_recording_headless, find_calibration_toml_path
 
 
@@ -43,13 +43,13 @@ def total_body_center_of_mass_file_path():
 
 
 @pytest.fixture
-def image_tracking_data_file_name():
+def image_tracking_data_file_path():
     return get_image_tracking_data_file_name(get_output_data_folder_path(pytest.sample_session_folder_path))
 
 
 @pytest.fixture
-def reprojection_error_file_name():
-    return get_reprojection_error_file_name(get_output_data_folder_path(pytest.sample_session_folder_path))
+def reprojection_error_file_path():
+    return get_reprojection_error_file_path(get_output_data_folder_path(pytest.sample_session_folder_path))
 
 if __name__ == "__main__":
     pytest_sessionstart()
