+++ freemocap/tests/test_mediapipe_skeleton_data_shape.py
@@ -11,7 +11,7 @@ from freemocap.utilities.get_number_of_frames_of_videos_in_a_folder import (
 def test_mediapipe_skeleton_data_shape(
     synchronized_video_folder_path: Union[str, Path],
     raw_skeleton_npy_file_path: Union[str, Path],
-    reprojection_error_file_name: Union[str, Path],
+        reprojection_error_file_path,
 ):
     """
     test that the `mediapipe 3d detection` process worked correctly by checking:
@@ -33,16 +33,16 @@ def test_mediapipe_skeleton_data_shape(
     ), f"3d skeleton data file should have 3 dimensions -  {raw_skeleton_npy_file_path}"
 
     assert Path(
-        reprojection_error_file_name
-    ).is_file(), f"3d skeleton reprojection error data file does not exist at {reprojection_error_file_name}"
+        reprojection_error_file_path
+    ).is_file(), f"3d skeleton reprojection error data file does not exist at {reprojection_error_file_path}"
 
     skeleton_reprojection_error_fr_mar = np.load(
-        reprojection_error_file_name
+        reprojection_error_file_path
     )
 
     assert (
         len(skeleton_reprojection_error_fr_mar.shape) == 2
-    ), f"3d skeleton reprojection error data file should have 2 dimensions {reprojection_error_file_name}"
+    ), f"3d skeleton reprojection error data file should have 2 dimensions {reprojection_error_file_path}"
 
     frame_count = get_number_of_frames_of_videos_in_a_folder(synchronized_video_folder_path)
     assert (
